22/8-07

+++ contest/grp/806.cpp
@@ -1,6 +1,6 @@
 #include <cstdio>
 
-int cnt['z'+1];
+long long cnt['z'+1];
 
 int main() {
   int n;
